AscnPROJ/TP24-25-main/TP24-25-main/checkpoints/checkpoints2_3/GrupoTP-13-master/docker/moonshot/requirements.txt
cryptography==3.4.6
Django==5.0.7
django-cors-headers==3.7.0
django-countries==7.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
luhn==0.2.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
PyJWT==1.7.1
python-dotenv==0.15.0
segno==1.3.1